refactor: split framework plugin implementations
