+++ packages/postcss/src/branches/style.ts
@@ -1,4 +1,9 @@
+import type { Result } from 'postcss'
 import type { IStyleHandlerOptions } from '../types'
+import { postprocessMiniProgramCss } from './mini-program'
+import { postprocessUniAppXUvueCss } from './uni-app-x-css-uvue'
+import { postprocessUniAppXWebviewCss } from './uni-app-x-css-webview'
+import { postprocessWebCss } from './web'
 
 export type PostcssAppType
   = | 'kbone'
@@ -14,9 +19,14 @@ export type PostcssAppType
 export type PostcssStyleBranch
   = | 'generic'
     | 'mini-program'
-    | 'uni-app-x-uvue'
-    | 'uni-app-x-webview'
+    | 'uni-app-x-css-uvue'
+    | 'uni-app-x-css-webview'
     | 'web'
+
+export interface PostcssStyleBranchProfile {
+  branch: PostcssStyleBranch
+  postprocess: (result: Result, options: IStyleHandlerOptions) => Result
+}
 
 export interface ResolvePostcssStyleBranchOptions extends Pick<IStyleHandlerOptions, 'platform' | 'uniAppX' | 'uniAppXCssTarget'> {
   appType?: PostcssAppType | undefined
@@ -44,7 +54,7 @@ function isMiniProgramAppType(appType: PostcssAppType | undefined) {
 
 export function resolvePostcssStyleBranch(options: ResolvePostcssStyleBranchOptions): PostcssStyleBranch {
   if (options.uniAppX) {
-    return options.uniAppXCssTarget === 'uvue' ? 'uni-app-x-uvue' : 'uni-app-x-webview'
+    return options.uniAppXCssTarget === 'uvue' ? 'uni-app-x-css-uvue' : 'uni-app-x-css-webview'
   }
   if (isWebLikePlatform(options.platform)) {
     return 'web'
@@ -55,10 +65,34 @@ export function resolvePostcssStyleBranch(options: ResolvePostcssStyleBranchOpti
   return 'generic'
 }
 
-export function shouldApplyUniAppXBaseCompatibility(branch: PostcssStyleBranch) {
-  return branch === 'uni-app-x-webview' || branch === 'uni-app-x-uvue'
-}
-
-export function shouldApplyUniAppXUvueCompatibility(branch: PostcssStyleBranch) {
-  return branch === 'uni-app-x-uvue'
+export function resolvePostcssStyleBranchProfile(options: ResolvePostcssStyleBranchOptions): PostcssStyleBranchProfile {
+  const branch = resolvePostcssStyleBranch(options)
+  switch (branch) {
+    case 'mini-program':
+      return {
+        branch,
+        postprocess: postprocessMiniProgramCss,
+      }
+    case 'uni-app-x-css-uvue':
+      return {
+        branch,
+        postprocess: postprocessUniAppXUvueCss,
+      }
+    case 'uni-app-x-css-webview':
+      return {
+        branch,
+        postprocess: postprocessUniAppXWebviewCss,
+      }
+    case 'web':
+      return {
+        branch,
+        postprocess: postprocessWebCss,
+      }
+    case 'generic':
+    default:
+      return {
+        branch,
+        postprocess: postprocessMiniProgramCss,
+      }
+  }
 }
