+++ packages/postcss/src/branches/uni-app-x-css-webview/index.ts
@@ -1,0 +1,7 @@
+import type { Result } from 'postcss'
+import type { IStyleHandlerOptions } from '../../types'
+import { applyUniAppXBaseCompatibility } from '../../compat/uni-app-x'
+
+export function postprocessUniAppXWebviewCss(result: Result, options: IStyleHandlerOptions) {
+  return applyUniAppXBaseCompatibility(result, options)
+}
