+++ packages/postcss/src/branches/web/index.ts
@@ -1,0 +1,6 @@
+import type { Result } from 'postcss'
+import type { IStyleHandlerOptions } from '../../types'
+
+export function postprocessWebCss(result: Result, _options: IStyleHandlerOptions) {
+  return result
+}
