+++ packages/postcss/src/index.ts
@@ -2,10 +2,10 @@
 export {
   type PostcssAppType,
   type PostcssStyleBranch,
+  type PostcssStyleBranchProfile,
   resolvePostcssStyleBranch,
   type ResolvePostcssStyleBranchOptions,
-  shouldApplyUniAppXBaseCompatibility,
-  shouldApplyUniAppXUvueCompatibility,
+  resolvePostcssStyleBranchProfile,
 } from './branches'
 export {
   type DynamicColorMixAlphaProtection,
