+++ packages/weapp-tailwindcss/src/bundlers/gulp/frameworks/native/index.ts
@@ -1,0 +1,6 @@
+import type { UserDefinedOptions } from '@/types'
+import { createNativeGulpPlugins } from '../../shared/create-native-framework-plugins'
+
+export function createNativeGulpFrameworkPlugins(options: UserDefinedOptions = {}) {
+  return createNativeGulpPlugins(options)
+}
