+++ packages/weapp-tailwindcss/src/bundlers/vite/frameworks/generic/index.ts
@@ -1,0 +1,8 @@
+import type { InternalUserDefinedOptions, UserDefinedOptions } from '@/types'
+import { createViteFrameworkPlugins } from '../../shared/create-framework-plugins'
+
+export function createGenericVitePlugins(options: UserDefinedOptions | InternalUserDefinedOptions = {}) {
+  return createViteFrameworkPlugins(options, {
+    branchName: 'generic-vite',
+  })
+}
