+++ packages/weapp-tailwindcss/src/bundlers/vite/frameworks/uni-app-x/index.ts
@@ -1,0 +1,24 @@
+import type { InternalUserDefinedOptions, UserDefinedOptions } from '@/types'
+import { createUniAppXPlugins } from '@/uni-app-x/vite'
+import { createViteFrameworkPlugins } from '../../shared/create-framework-plugins'
+
+export function createUniAppXVitePlugins(options: UserDefinedOptions | InternalUserDefinedOptions = {}) {
+  return createViteFrameworkPlugins(options, {
+    branchName: 'uni-app-x-vite',
+    createExtraPlugins: context => createUniAppXPlugins({
+      appType: 'uni-app-x',
+      customAttributesEntities: context.customAttributesEntities,
+      disabledDefaultTemplateHandler: context.disabledDefaultTemplateHandler,
+      ensureRuntimeClassSet: context.ensureRuntimeClassSet,
+      generateCss: context.generateCss,
+      getResolvedConfig: context.getResolvedConfig,
+      isEnabled: context.isEnabled,
+      isIosPlatform: context.isIosPlatform,
+      jsHandler: context.jsHandler,
+      mainCssChunkMatcher: context.mainCssChunkMatcher,
+      runtimeState: context.runtimeState,
+      styleHandler: context.styleHandler,
+      uniAppX: context.uniAppX,
+    }),
+  })
+}
