+++ packages/weapp-tailwindcss/src/bundlers/vite/frameworks/uni-app/index.ts
@@ -1,0 +1,8 @@
+import type { InternalUserDefinedOptions, UserDefinedOptions } from '@/types'
+import { createViteFrameworkPlugins } from '../../shared/create-framework-plugins'
+
+export function createUniAppVitePlugins(options: UserDefinedOptions | InternalUserDefinedOptions = {}) {
+  return createViteFrameworkPlugins(options, {
+    branchName: 'uni-app-vite',
+  })
+}
