+++ packages/weapp-tailwindcss/src/bundlers/vite/frameworks/weapp-vite/index.ts
@@ -1,0 +1,8 @@
+import type { InternalUserDefinedOptions, UserDefinedOptions } from '@/types'
+import { createViteFrameworkPlugins } from '../../shared/create-framework-plugins'
+
+export function createWeappVitePlugins(options: UserDefinedOptions | InternalUserDefinedOptions = {}) {
+  return createViteFrameworkPlugins(options, {
+    branchName: 'weapp-vite',
+  })
+}
