+++ packages/weapp-tailwindcss/src/bundlers/webpack/frameworks/taro/index.ts
@@ -1,0 +1,10 @@
+import type { InternalUserDefinedOptions, UserDefinedOptions } from '@/types'
+import { WebpackFrameworkPlugin } from '../../shared/create-framework-plugin'
+
+export class TaroWebpackPlugin extends WebpackFrameworkPlugin {
+  constructor(options: UserDefinedOptions | InternalUserDefinedOptions = {}) {
+    super(options, {
+      branchName: 'taro-webpack',
+    })
+  }
+}
