+++ packages/weapp-tailwindcss/src/bundlers/webpack/frameworks/uni-app/index.ts
@@ -1,0 +1,10 @@
+import type { InternalUserDefinedOptions, UserDefinedOptions } from '@/types'
+import { WebpackFrameworkPlugin } from '../../shared/create-framework-plugin'
+
+export class UniAppWebpackPlugin extends WebpackFrameworkPlugin {
+  constructor(options: UserDefinedOptions | InternalUserDefinedOptions = {}) {
+    super(options, {
+      branchName: 'uni-app-webpack',
+    })
+  }
+}
